import/requirements.txt
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.22
GeoAlchemy2==0.12.5
geopandas==0.10.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
load-dotenv==0.1.0
munch==2.5.0
numpy==1.21.6
packaging==22.0
pandas==1.3.5
psycopg2==2.9.5
py4j==0.10.9.5
pycodestyle==2.10.0
pyproj==3.2.1
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
shapely==2.0.0
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
